Merge Sangoma updates into master
This is the culmination of the Sangoma fork work to update
react-native-callkeep to support more consistent SDK features across
both iOS and Android. Specifically, adding multi-call tracking to
Android and a more consistent JS interface. This is a major version
bump, as updating to this will break existing applications. We've
attempted to keep the README updated with the proper instructions,
although we did *not* write a transition guide. Happy to help contribute
additional work if desired prior to merging, or to be a contributor on
the community project to help maintain moving forward and avoid our own
fork.