Add a troubleshooting note on using a valid uuid
Was stuck for some time trying to figure out why the incoming call wasn't being received and it turned out, it was simply because I had supplied an invalid `uuid`.  The `didDisplayIncomingCall` method also returns without any errors, so there's no explanation for why the screen didn't show up. Might be great if the library threw an error in `didDisplayIncomingCall` if an invalid `uuid` is supplied.